Moved wip-check back to pr-checks.yml.
It turns out githubs composite actions are not actually
compositions of actions but compositions of shell commands
i.e. they are more of a script action then a composition
action. But they don't support typescript, the actions
toolkit etc. so we would not to create a full standalon
actions which is like the 80% github-script command.